:goal_net: error handling getting user by id

+++ src/api/controllers/users.controller.ts
@@ -1,6 +1,8 @@
 import httpStatus from 'http-status';
 import { NextFunction, Request, Response } from 'express';
+import { APIError } from '@rcebrian/tfg-rcebrian-common';
 import { User } from '../repository/mysql/mysql.repository';
+import logger from '../../config/winston.config';
 
 /**
  * Get all user from database
@@ -13,6 +15,7 @@ export const findAllUsers = (req: Request, res: Response, next: NextFunction) =>
     .then((data) => {
       res.status(httpStatus.OK).json({ data });
     }).catch((err) => {
+      logger.error(err);
       next(err);
     });
 };
@@ -28,8 +31,17 @@ export const findById = (req: Request, res: Response, next: NextFunction) => {
 
   User.findOne({ where: { id: userId } })
     .then((data) => {
-      res.status(httpStatus.OK).json({ data });
+      if (data) {
+        res.status(httpStatus.OK).json({ data });
+      } else {
+        throw new APIError({
+          message: 'Not found',
+          status: httpStatus.NOT_FOUND,
+          stack: `Can't get user. User [${userId}] not exists`,
+        });
+      }
     }).catch((err) => {
+      logger.error(err.stack);
       next(err);
     });
 };
@@ -56,6 +68,7 @@ export const update = (req: Request, res: Response, next: NextFunction) => {
     .then(() => {
       res.status(httpStatus.ACCEPTED).json();
     }).catch((err) => {
+      logger.error(err);
       next(err);
     });
 };
@@ -68,10 +81,25 @@ export const update = (req: Request, res: Response, next: NextFunction) => {
 export const remove = (req: Request, res: Response, next: NextFunction) => {
   const { userId } = req.params;
 
-  User.destroy({ where: { id: userId } })
-    .then((data) => {
-      res.status(httpStatus.NO_CONTENT).json({ data });
-    }).catch((err) => {
-      next(err);
-    });
+  User.findOne(
+    { where: { id: userId } },
+  ).then((user) => {
+    if (!user) {
+      throw new APIError({
+        message: 'Not found',
+        status: httpStatus.NOT_FOUND,
+        stack: `Can't delete user. User [${userId}] not exists`,
+      });
+    }
+    User.destroy({ where: { id: userId } })
+      .then((data) => {
+        res.status(httpStatus.NO_CONTENT).json({ data });
+      }).catch((err) => {
+        logger.error(err.stack);
+        next(err);
+      });
+  }).catch((err) => {
+    logger.error(err.stack);
+    next(err);
+  });
 };
